Fix code smells suggested by sonarqube part - 9 (#2583)
* Fix security hotspot suggested by sonarqube

* Fix all remaining medium and low security hotspots

* Fix remaining security hotspots

* https://jenkins.dev.verta.ai/job/build/job/autoformat/job/modeldb-common/114/

* Fix code smells suggested by sonarqube part - 1

* Fix code smells suggested by sonarqube part - 2

* Fix code smells suggested by sonarqube part - 3

* Fix code smells suggested by sonarqube part - 4

* Fix code smells suggested by sonarqube part - 5

* Fix code smells suggested by sonarqube part - 6

* Fix code smells suggested by sonarqube part - 7 (`var` changes)

* Rename Objects based on sonarqube suggestions - Sonar changes part 8

* https://jenkins.dev.verta.ai/job/build/job/autoformat/job/modeldb-backend/157/

* revert pom changes

* Fix code smells suggested by sonarqube part - 9

* fix pom for sonarqube

* https://jenkins.dev.verta.ai/job/build/job/autoformat/job/modeldb-common/159/

* fix var at missing places

* Fix config parameter name to support existing config fields

* Fix review comments

* Removed test file from prs

+++ backend/src/main/java/ai/verta/modeldb/utils/RdbmsUtils.java
@@ -2172,7 +2172,8 @@ public class RdbmsUtils {
       Object value,
       Map<String, Object> parametersMap) {
     @SuppressWarnings("squid:S2245")
-    long timestamp = index0 + new Random(System.nanoTime()).nextInt(Integer.MAX_VALUE);
+    long timestamp =
+        (long) index0 + (long) new Random(System.nanoTime()).nextInt(Integer.MAX_VALUE);
     String key;
     switch (operator.ordinal()) {
       case OperatorEnum.Operator.GT_VALUE:
